forcibly change NLS_DATE_FORMAT to 'YYYYMMDD' for oracle flavors
RStudio complained about 'java.sql.SQLDataException: ORA-01861:
literal does not match format string' and could only get 44%
through the initial Achilles run.

The underlying problem was tied to the use of CAST on an Oracle
datefield in the SQL command shown in the errorReport.txt file.
CASE is good because it is ANSI standard sql but it presupposes
the date format is 'YYYYMMDD'.