Restricting to observation period
As the CDM now allows records outside the observation period - this messes up some of the achilles plots.

I edited the data dashboard and visit analyses to restrict to things that overlap with an observation period.

for analysis 211 - I restricted to visits contained within an observation period rather than overlapping